requirement.txt
openai==1.97.1
python-dotenv==1.1.1
pydantic==2.11.7
pandas==2.3.1
openpyxl==3.1.5
tqdm==4.67.1